+++ sql/tbl_community.sql
@@ -1,0 +1,45 @@
+INSERT INTO tbl_community (category, content, created_at, like_count, report_count, title, updated_at, view_count, member_id)
+VALUES
+    ('FRIENDS', '이번 주 영화 모임 참여자 모집합니다.', '2023-07-06', 12, 0, '영화 모임', '2023-07-06', 110, 1),
+    ('SHARE', '좋은 책을 추천합니다.', '2023-07-08', 22, 0, '책 추천', '2023-07-08', 190, 1),
+    ('FRIENDS', '온라인 게임 같이 할 사람 구합니다.', '2023-07-10', 6, 0, '게임 친구 모집', '2023-07-10', 120, 1),
+    ('SHARE', '유용한 웹사이트를 소개합니다.', '2023-07-12', 13, 0, '웹사이트 추천', '2023-07-12', 170, 1),
+    ('FREE', '오늘 먹은 점심 메뉴는 정말 맛있었어요.', '2023-07-13', 9, 0, '점심 이야기', '2023-07-13', 160, 1),
+    ('FRIENDS', '함께 여행 가실 분 구합니다.', '2023-07-14', 14, 0, '여행 친구 모집', '2023-07-14', 150, 1),
+    ('NOTICE', '이벤트 참여 방법 안내', '2023-07-15', 19, 0, '이벤트 안내', '2023-07-15', 220, 1),
+    ('SHARE', '프로그래밍 관련 팁 공유합니다.', '2023-07-16', 16, 0, '프로그래밍 팁', '2023-07-16', 200, 1),
+    ('FREE', '오늘 저녁에 맛있는 요리 해먹었어요.', '2023-07-17', 11, 0, '요리 이야기', '2023-07-17', 140, 1),
+    ('FRIENDS', '운동 같이 하실 분 구합니다.', '2023-07-18', 7, 0, '운동 친구 모집', '2023-07-18', 130, 1),
+    ('NOTICE', '정기 점검 안내', '2023-07-19', 21, 1, '정기 점검', '2023-07-19', 210, 1),
+    ('SHARE', '좋은 앱을 추천합니다.', '2023-07-20', 17, 0, '앱 추천', '2023-07-20', 180, 1),
+    ('FREE', '오늘 하루 어떻게 보내셨나요?', '2023-07-21', 12, 0, '하루 이야기', '2023-07-21', 170, 1),
+    ('FRIENDS', '영화 같이 보실 분 구합니다.', '2023-07-22', 15, 0, '영화 친구 모집', '2023-07-22', 160, 1),
+    ('SHARE', '간단한 요리 레시피를 공유합니다.', '2023-07-24', 14, 0, '요리 레시피', '2023-07-24', 200, 1),
+    ('FREE', '오늘의 명언: "행복은 마음 속에 있다."', '2023-07-25', 18, 0, '명언', '2023-07-25', 150, 1),
+    ('FRIENDS', '주말 등산 같이 하실 분?', '2023-07-26', 13, 0, '등산 친구 모집', '2023-07-26', 140, 1),
+    ('NOTICE', '회원 등급 변경 안내', '2023-07-27', 22, 1, '등급 변경', '2023-07-27', 230, 1),
+    ('SHARE', '유용한 학습 자료를 공유합니다.', '2023-07-28', 20, 0, '학습 자료', '2023-07-28', 190, 1),
+    ('FREE', '오늘은 정말 좋은 날입니다.', '2023-07-29', 10, 0, '좋은 날', '2023-07-29', 170, 1),
+    ('FRIENDS', '음악 감상 모임에 참여하세요.', '2023-07-30', 9, 0, '음악 모임', '2023-07-30', 120, 1),
+    ('FREE', '야오야오', '2024-08-17', null, null, '커뉴미ㅣ니', '2024-08-17', null, 1),
+    ('FREE', '가나라ㅏㄷ마ㅏ다', '2024-08-17', null, null, '제목은 제목', '2024-08-17', null, 1),
+    ('FRIENDS', '야호', '2024-08-17', null, null, '강쥐 친구 찾아유', '2024-08-17', null, 1),
+    ('SHARE', '나눠요 나눠 우리모두 야호', '2024-08-17', null, null, '나눔 파리 야호', '2024-08-17', null, 1),
+    ('FREE', '자융', '2024-08-17', null, null, '자유를 찾아서', '2024-08-17', null, 1),
+    ('FREE', '야옹옹', '2024-08-17', null, null, 'ㅇ호야홍', '2024-08-17', null, 1),
+    ('SHARE', '야호', '2024-08-17', null, null, '나눠볼까유', '2024-08-17', null, 1),
+    ('SHARE', '11111', '2024-08-17', null, null, '야웅웅', '2024-08-17', null, 1),
+    ('SHARE', '나눔', '2024-08-17', null, null, '나문뉸', '2024-08-17', null, 1),
+    ('FRIENDS', '챠오옹', '2024-08-17', null, null, '나는친구 찾기', '2024-08-17', null, 1),
+    ('FREE', '야호', '2024-08-17', null, null, '몇번째 커뮤니티', '2024-08-17', null, 1),
+    ('SHARE', '야호', '2024-08-17', null, null, '나눔 커뉴니니니', '2024-08-17', null, 1),
+    ('SHARE', '야호', '2024-08-17', null, null, '나눠볼까?', '2024-08-17', null, 1),
+    ('FREE', 'ㅇㅇㅇㅇㅇㅇㅇ', '2024-08-17', null, null, '44번째', '2024-08-17', null, 1),
+    ('FRIENDS', 'ㅇ', '2024-08-17', null, null, '강친모', '2024-08-17', null, 1),
+    ('FRIENDS', 'ㅇ', '2024-08-17', null, null, '강아지 친구 찾기 6시 54분', '2024-08-17', null, 1),
+    ('SHARE', 'ddd', '2024-08-17', null, null, 'ddd', '2024-08-17', null, 1),
+    ('FREE', '2', '2024-08-17', null, null, '자유 7시 2붖', '2024-08-17', null, 1),
+    ('SHARE', '야오야오ㅑㅇㅇ', '2024-08-17', null, null, '야호야호야ㅗ양', '2024-08-17', null, 1),
+    ('SHARE', '4', '2024-08-17', null, null, '나눔 7시 6분', '2024-08-17', null, 1),
+    ('FRIENDS', 's', '2024-08-17', null, null, '강아지 친귱 7 시 14', '2024-08-17', null, 1),
+    ('SHARE', 'dddd', '2024-08-17', null, null, 'h@h 7:17', '2024-08-17', null, 1);
