+++ src/main/java/com/sh/pettopia/Hojji/user/member/dto/MemberRegistRequestDto.java
@@ -43,7 +43,6 @@ public class MemberRegistRequestDto {
 
     // 생년 월일
     private LocalDate birth;
-    private Timestamp timestamp;
 
     // 주소
     private String address;
