+++ src/main/resources/templates/community/postDetail.html
@@ -274,7 +274,7 @@
 
         function deletePost ()  {
             console.log('deletePost 함수 호출');
-            confirm('정말 이 메뉴를 삭제하시겠습니까?')
+            confirm('정말 이 게시글을 삭제하시겠습니까?')
             && document.deletePostFrm.submit();
         }
